requirements.txt
sounddevice==0.4.6
numpy>=1.26.0
openai>=1.12.0
openai-whisper>=20231117
streamlit>=1.32.0
torch>=2.2.0
torchaudio>=2.2.0
ffmpeg-python>=0.2.0 